# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Info": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"pages": {
						"type": "number",
						"format": "double"
					},
					"next": {
						"type": "string",
						"nullable": true
					},
					"prev": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"count",
					"pages",
					"next",
					"prev"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Episode": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"air_date": {
						"type": "string"
					},
					"episode": {
						"type": "string"
					},
					"characters": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					},
					"created": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"air_date",
					"episode",
					"characters",
					"url",
					"created"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EpisodesResponseDTO": {
				"properties": {
					"info": {
						"$ref": "#/components/schemas/Info"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/Episode"
						},
						"type": "array"
					}
				},
				"required": [
					"info",
					"results"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pick_Episode.Exclude_keyofEpisode.characters__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"air_date": {
						"type": "string"
					},
					"episode": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"created": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"air_date",
					"episode",
					"url",
					"created"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Episode.characters_": {
				"$ref": "#/components/schemas/Pick_Episode.Exclude_keyofEpisode.characters__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Character": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"species": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"origin": {
						"properties": {
							"url": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"url",
							"name"
						],
						"type": "object"
					},
					"location": {
						"properties": {
							"url": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"url",
							"name"
						],
						"type": "object"
					},
					"image": {
						"type": "string"
					},
					"episode": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					},
					"created": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"status",
					"species",
					"type",
					"gender",
					"origin",
					"location",
					"image",
					"episode",
					"url",
					"created"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GetCharactersByEpisodeIdResponseDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_Episode.characters_"
					},
					{
						"properties": {
							"characters": {
								"items": {
									"$ref": "#/components/schemas/Character"
								},
								"type": "array"
							}
						},
						"required": [
							"characters"
						],
						"type": "object"
					}
				]
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/episodes": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EpisodesResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "episode",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/episodes/{id}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Episode"
								}
							}
						}
					}
				},
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/episodes/{id}/characters": {
			"get": {
				"operationId": "GetCharactersByEpisodeId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCharactersByEpisodeIdResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Episodes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}